11.27

+++ hw/syn/xilinx/dut/common.mk
@@ -31,16 +31,21 @@ project_1/sources.txt:
 build: $(PROJECT).xpr
 $(PROJECT).xpr: project_1/sources.txt
 ifdef FPU_IP
-	MAX_JOBS=$(JOBS) FPU_IP=project_1/ip $(VIVADO) -mode batch -source $(SRC_DIR)/project.tcl -tclargs $(TOP_LEVEL_ENTITY) $(DEVICE) project_1/sources.txt $(SRC_DIR)/project.xdc $(SCRIPT_DIR)
+	MAX_JOBS=$(JOBS) FPU_IP=project_1/ip TOOL_DIR=$(SCRIPT_DIR) $(VIVADO) -mode batch -source $(SRC_DIR)/project.tcl -tclargs $(TOP_LEVEL_ENTITY) $(DEVICE) project_1/sources.txt $(SRC_DIR)/project.xdc
 else
-	MAX_JOBS=$(JOBS) $(VIVADO) -mode batch -source $(SRC_DIR)/project.tcl -tclargs $(TOP_LEVEL_ENTITY) $(DEVICE) project_1/sources.txt $(SRC_DIR)/project.xdc $(SCRIPT_DIR)
+	MAX_JOBS=$(JOBS) TOOL_DIR=$(SCRIPT_DIR) $(VIVADO) -mode batch -source $(SRC_DIR)/project.tcl -tclargs $(TOP_LEVEL_ENTITY) $(DEVICE) project_1/sources.txt $(SRC_DIR)/project.xdc
 endif
 
 clean:
+ifndef RESUME
 	rm -rf project_1
 	rm -rf .Xil
 	rm -f *.rpt
-	rm -f vivado*.log
-	rm -f vivado*.jou
+	rm -f *.log
+	rm -f *.jou
+	rm -f *.dcp
+else
+	@echo "RESUME is defined, skipping clean."
+endif
 
 .PHONY: all gen-sources build clean
